+++ packages/js/admin-e2e-tests/src/pages/PaymentsSetup.ts
@@ -25,6 +25,15 @@ export class PaymentsSetup extends BasePage {
 		await this.clickButtonWithText( 'Got it' );
 	}
 
+	async showOtherPaymentMethods(): Promise< void > {
+		const selector = '.woocommerce-task-payments button.toggle-button';
+		await this.page.waitForSelector( selector );
+		const toggleButton = await this.page.$(
+			`${ selector }[aria-expanded=false]`
+		);
+		await toggleButton?.click();
+	}
+
 	async goToPaymentMethodSetup(
 		method: PaymentMethodWithSetupButton
 	): Promise< void > {
@@ -39,14 +48,6 @@ export class PaymentsSetup extends BasePage {
 		} else {
 			await button.click();
 		}
-	}
-
-	async methodHasBeenSetup( method: PaymentMethod ): Promise< void > {
-		const selector = `.woocommerce-task-payment-${ method }`;
-		await this.page.waitForSelector( selector );
-		expect(
-			await getElementByText( '*', 'Manage', selector )
-		).toBeDefined();
 	}
 
 	async enableCashOnDelivery(): Promise< void > {