+++ packages/js/admin-e2e-tests/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+-   Add E2E tests to disabled welcome modal #32505
+
+-   Update test for payment task. #32467
+
 # 1.0.0
 
 -   Add returned type annotations and remove unused vars. #8020