+++ packages/js/data/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+-   Update dependency `@wordpress/hooks` to ^3.5.0
+-   Add `is_offline` attribute for `Plugin` type. #32467
+
 # 3.1.0
 
 -   Add "moment" to peerDependencies. #8349